chore(attendance): move calendar UI composables from ScheduleScreen to ui components
